requirements.txt
Django==1.4.5
Jinja2==2.7
MarkupSafe==0.18
Pillow==2.1.0
Pygments==1.6
South==0.8.1
Sphinx==1.2b1
argparse==1.2.1
coverage==3.6
django-allauth==0.12.0
django-avatar==2.0a10
docutils==0.11
httplib2==0.8
oauth2==1.5.211
oauthlib==0.5.1
python-twitter==1.0
requests==1.2.3
requests-oauthlib==0.3.2
simplejson==3.3.0
wsgiref==0.1.2
